Don't retrieve all EC2 images when only one is needed
Previously all images were retrieved from the provider when deploying a node.
The image list is not created by Libcloud in an efficient way, which caused the
libcloud_wrapper process to use ~600 MB memory after the "list_images" call in
case of Amazon EC2.

This problem is described in more detail in the following pages:

- https://forge.univention.org/bugzilla/show_bug.cgi?id=35977
- https://issues.apache.org/jira/browse/LIBCLOUD-294

The first page says the followings:

- "For the purpose of instance creation we load all available images via
  libcloud and cache it in the local memory. Just receiving this information
  uses 400-500MiB of RAM."
- "libcloud internally uses the ElementTree XML API to parse the large response
  from AWS. If the most efficient C implementation "lxml.etree" from
  "pyhton-lxml" is not available, it falls back to the Python internal
  xml.etree.ElementTree implementation, which seems to be a memory hog"

The solution is to specify a filter when the provider is EC2. (The OpenStack
provider doesn't have such a filter; the get_images() function could be used in
that case.)